Restructures interfaces
