+++ composables/useContent.ts
@@ -1,45 +1,4 @@
-export interface DeviceSpecificHeightType {
-  mobile: string;
-  tablet: string;
-  desktop: string;
-}
-
-export interface Image {
-  url: string;
-  alt_text: string;
-  focal_point: string | undefined;
-  name: string;
-  title: string;
-}
-
-export interface Background {
-  image: Image;
-  color: string;
-}
-
-export interface MediaArguments {
-  maxHeight?: number | undefined;
-  maxWidth?: number | undefined;
-  skipAutoFormat?: boolean | undefined;
-  focalPoint?: string | undefined;
-}
-
-export interface MenuDepartment {
-  key: string;
-  title: string;
-  slug: string;
-  target: string;
-  image: Image;
-  sub_departments: MenuDepartment[];
-}
-
-export interface HeaderMenu {
-  departments: MenuDepartment[];
-}
-
-export interface FooterMenu {
-  departments: MenuDepartment[];
-}
+import { MediaArguments } from '~/interfaces/Content';
 
 export const useMediaHandler = () => {
   function toImageUrl(url: string, {
