+++ interfaces/page.ts
@@ -1,0 +1,5 @@
+export interface PageMeta {
+  title: string;
+  description: string;
+  seamless_header: boolean;
+}
